+++ Source/Frontend/UI/UI.csproj
@@ -375,6 +375,7 @@
       <DependentUpon>RTC_Test_Form.cs</DependentUpon>
     </Compile>
     <Compile Include="GameProtection.cs" />
+    <Compile Include="SortableBindingList.cs" />
     <Compile Include="Input\Bindings.cs" />
     <Compile Include="Input\Gamepad.cs" />
     <Compile Include="Input\Gamepad360.cs" />
